Change `syntactic_environment` to a non-template class
